+++ src/pages/Projects/components/RoleCard/RoleCard.tsx
@@ -15,30 +15,6 @@ interface RoleCardProps {
   handleDeleteRole: (role: Role) => void;
   handleEditRole: (role: Role) => void;
 }
-const mockRole = {
-  id: "5f6a2f57-2f0e-48d7-ac28-1aae14792e57",
-  startDate: "Mon Aug 22 2022 00:00:00 GMT-0400 (Eastern Daylight Time)",
-  startConfidence: 0.5,
-  endDate: "Fri Oct 14 2022 00:00:00 GMT-0400 (Eastern Daylight Time)",
-  endConfidence: 0.6,
-  projectId: "f9061899-80af-4337-ae32-913edf0f6fed",
-  assignments: [
-    {
-      id: "2f313b76-26c9-4dc1-8708-e964c3ce4b6f",
-      employeeId: "0a2b5826-5a56-4f4f-90f6-b8dfbeae6598",
-      roleId: "5f6a2f57-2f0e-48d7-ac28-1aae14792e57",
-      startDate: "Thu Apr 04 2024 20:00:00 GMT-0400 (Eastern Daylight Time)",
-      endDate: "Sun Dec 22 2024 19:00:00 GMT-0500 (Eastern Standard Time)",
-      employee: {
-        id: "0a2b5826-5a56-4f4f-90f6-b8dfbeae6598",
-        name: "Celestino Reilly",
-        startDate: "Sun Aug 21 2022 20:00:00 GMT-0400 (Eastern Daylight Time)",
-        endDate: null,
-      },
-    },
-  ],
-  skills: ['{id: "a8ff7e2b-6097-4914-91a5-7b7ca704eacc", name: …}'],
-};
 
 export default function RoleCard({
   role,
@@ -206,7 +182,7 @@ export default function RoleCard({
               </Text>
             </Td>
             <Td colSpan={6}>
-              <DataGanttLine timeline={timeline} role={mockRole} />
+              <DataGanttLine timeline={timeline} role={role} />
             </Td>
           </Tr>
           <Tr>
