+++ src/services/helpers/gantt/ganttCell/components/SplitGantt/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./SplitGantt";
+export * from "./SplitGantt";
